career/serializers.py
from rest_framework import serializers
from .models import JobApplication,InternshipApplication,Certificate

class JobApplicationSerializer(serializers.ModelSerializer):
    class Meta:
        model = JobApplication
        fields = "__all__"

class InternshipApplicationSerializer(serializers.ModelSerializer):
    class Meta:
        model = InternshipApplication
        fields = "__all__"

class CertificateSerializer(serializers.ModelSerializer):
    class Meta:
        model = Certificate
        fields = ['id','certificateno','name','role','dateofissue','dateofexpiry','status']