requirements.txt
asgiref==3.2.10
Django==3.0.8
django-crispy-forms==1.9.1
django-summernote==0.8.11.6
django-taggit==1.3.0
djangorestframework==3.11.0
gunicorn==20.0.4
Pillow==7.2.0
pytz==2020.1
sqlparse==0.3.1
